// src/app/Pages/Models/bannermodal.ts
export class bannermodal {
  ID: number = 0;
  TITLE: any;
  TITLE_COLOR: string = '';
  CUSTOMER_TYPE: any = 'BO';
  SUB_TITLE: any;
  SUB_TITLE_COLOR: string = '';
  BANNER_TYPE: any = 'M';
  BANNER_FOR: any = 'W';
  SUB_TITLE_1: string = '';
  SUB_TITLE_COLOR_1: any;
  IS_FOR_SHOP: boolean = false;
  STATUS: boolean = true;
  SEQ_NO: number = 0;
  IMAGE_URL: any;
}